🚚 Rename: activity 공통 파일 common 폴더로 이동
