+++ src/features/activity/common/ui/skeleton/ActivityCardSkeleton.tsx
